# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  model.CreateMenuRequest:
    properties:
      categories:
        type: string
      name:
        type: string
      price:
        type: number
    required:
    - name
    - price
    type: object
  model.CreateMenuResponse:
    properties:
      categories:
        type: string
      id:
        type: integer
      name:
        type: string
      price:
        type: number
    type: object
...
  model.GetMenuResponse:
    $ref: '#/definitions/model.CreateMenuResponse'
...
  model.MenuResponse:
    properties:
      menu:
        type: object
    type: object
...
  model.UpdateMenuRequest:
    $ref: '#/definitions/model.CreateMenuRequest'
  model.UpdateMenuResponse:
    $ref: '#/definitions/model.CreateMenuResponse'
...
  web.ErrJSONResponse:
    properties:
      error:
        $ref: '#/definitions/web.Error'
        type: object
      process_time:
        type: string
      status:
        type: string
      success:
        example: false
        type: boolean
    type: object
  web.Error:
    properties:
      message:
        type: string
    type: object
  web.JSONResponse:
    properties:
      data:
        type: object
        x-omitempty: true
      process_time:
        type: string
      status:
        type: string
      success:
        example: true
        type: boolean
    type: object
host: localhost:9000
info:
  contact:
    email: support.family-catering@example.com
    name: Family catering API Support
    url: http://www.family-catering.com/support
  description: Documentation for Family Catering API.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Family Catering API
  version: "1.0"
paths:
...
  /menu:
    get:
      description: Show list of menus by (optionally) by given limit of offset
      parameters:
      - default: Bearer <your access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Pagination limit
        format: int64
        in: query
        name: limit
        type: integer
      - description: Pagination offset
        format: int64
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Ok
          schema:
            allOf:
            - $ref: '#/definitions/web.JSONResponse'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/model.MenuResponse'
                  - properties:
                      menu:
                        items:
                          $ref: '#/definitions/model.GetMenuResponse'
                        type: array
                    type: object
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/web.ErrJSONResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/web.ErrJSONResponse'
      summary: Show list of menus
      tags:
      - menu
    post:
      consumes:
      - application/json
      description: Create a new menu
      parameters:
      - default: Bearer <your access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: body request
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/model.CreateMenuRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Ok
          schema:
            allOf:
            - $ref: '#/definitions/web.JSONResponse'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/model.MenuResponse'
                  - properties:
                      menu:
                        $ref: '#/definitions/model.CreateMenuResponse'
                    type: object
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/web.ErrJSONResponse'
        "422":
          description: Unprocessable entity
          schema:
            $ref: '#/definitions/web.ErrJSONResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/web.ErrJSONResponse'
      summary: Create a menu
      tags:
      - menu
  /menu/{id}:
    delete:
      description: Delete menu by given id
      parameters:
      - description: Menu id
        format: int64
        in: path
        name: id
        required: true
        type: integer
      - default: Bearer <your access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Ok
          schema:
            $ref: '#/definitions/web.JSONResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/web.ErrJSONResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.ErrJSONResponse'
        "404":
          description: Menu not found
          schema:
            $ref: '#/definitions/web.ErrJSONResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/web.ErrJSONResponse'
      summary: Delete menu
      tags:
      - menu
    get:
      description: Show interest menu detail by given id
      parameters:
      - default: Bearer <your access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Menu id
        format: int64
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Ok
          schema:
            allOf:
            - $ref: '#/definitions/web.JSONResponse'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/model.MenuResponse'
                  - properties:
                      menu:
                        $ref: '#/definitions/model.GetMenuResponse'
                    type: object
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/web.ErrJSONResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.ErrJSONResponse'
        "404":
          description: Menu not found
          schema:
            $ref: '#/definitions/web.ErrJSONResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/web.ErrJSONResponse'
      summary: Get menu
      tags:
      - menu
    put:
      consumes:
      - application/json
      description: Update menu by given id
      parameters:
      - description: Menu id
        format: int64
        in: path
        name: id
        required: true
        type: integer
      - default: Bearer <your access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: body request
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/model.UpdateMenuRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Ok
          schema:
            allOf:
            - $ref: '#/definitions/web.JSONResponse'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/model.MenuResponse'
                  - properties:
                      menu:
                        $ref: '#/definitions/model.UpdateMenuResponse'
                    type: object
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/web.ErrJSONResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.ErrJSONResponse'
        "422":
          description: Unprocessable entity
          schema:
            $ref: '#/definitions/web.ErrJSONResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/web.ErrJSONResponse'
      summary: Update  menu
      tags:
      - menu
  /menu/{name}:
    get:
      description: Show interest menu detail by given name
      parameters:
      - default: Bearer <your access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Menu name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Ok
          schema:
            allOf:
            - $ref: '#/definitions/web.JSONResponse'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/model.MenuResponse'
                  - properties:
                      menu:
                        $ref: '#/definitions/model.GetMenuResponse'
                    type: object
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/web.ErrJSONResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.ErrJSONResponse'
        "404":
          description: Menu not found
          schema:
            $ref: '#/definitions/web.ErrJSONResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/web.ErrJSONResponse'
      summary: Get menu by given name
      tags:
      - menu
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"